By default, getID3 creates HTML version for all tags. In particular, it will convert the picture tag (which could be a bug). Depending on cover arts included in audio files, this feature triggers #354.