@lGuillaume124 2019-01-31
Could you complete the changelog and add a "Fixed" subtitle under the "Unreleased" section please?

Thank you for your contribution, I will try and merge it as soon as I can.