@lGuillaume124 2019-02-02
Awesome! One line changed and the parsing is more than ten times faster! Nice catch, thank you very much!

Without the patch:
```
$ time ./cake sonerezh sync --import --force
[...]
[INFO   ] Imported: 48 file(s).
./cake sonerezh sync --import --force  12,54s user 0,59s system 91% cpu 14,411 total
```

With the patch:
```
$ time ./cake sonerezh sync --import --force
[...]
[INFO   ] Imported: 48 file(s).
./cake sonerezh sync --import --force  0,72s user 0,36s system 65% cpu 1,655 total
```